Add last missing feature.
With this the base msc program is actually feature complete. More
features might follow in the future as needs change but for now the
program will enter the polish phase to fix everything wrong and broken
about this, which is a lot to a comical degree, my deepest apologies to
future me and anyone trying to read this hot pile of steaming garbage.